# api_go/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for SmartOrganizr.",
        "title": "SmartOrganizr API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/clubs/{clubId}/messages/candidates": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "List members who can be chatted with",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club ID",
                        "name": "clubId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ClubMessageCandidateDto"
                            }
                        }
                    }
                }
            }
        },
        "/v1/clubs/{clubId}/messages/chats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "List the user's chats in a club",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club ID",
                        "name": "clubId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ClubChatSummaryDto"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Start a new chat with a recipient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club ID",
                        "name": "clubId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New chat payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClubChatCreateDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ClubChatCreatedDto"
                        }
                    }
                }
            }
        },
        "/v1/clubs/{clubId}/messages/chats/{chatId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "List messages in a chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club ID",
                        "name": "clubId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chatId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ClubChatMessageDto"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Append a message to a chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club ID",
                        "name": "clubId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chatId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClubChatPostMessageDto"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/v1/concerts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "concerts"
                ],
                "summary": "List concerts for the current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ConcertDto"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "concerts"
                ],
                "summary": "Create a new concert",
                "parameters": [
                    {
                        "description": "Concert payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ConcertPostDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ConcertDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/concerts/{concertId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "concerts"
                ],
                "summary": "Get a single concert",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Concert ID",
                        "name": "concertId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ConcertDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "concerts"
                ],
                "summary": "Delete a concert",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Concert ID",
                        "name": "concertId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/events/{userId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "List events for a user since a given date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (must match authenticated user)",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Earliest event date (RFC3339)",
                        "name": "since",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Event"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.Author": {
            "type": "object",
            "required": [
                "extraInformation",
                "id",
                "name"
            ],
            "properties": {
                "extraInformation": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.ClubChatCreateDto": {
            "type": "object",
            "required": [
                "content",
                "recipient_user_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "recipient_user_id": {
                    "type": "string"
                }
            }
        },
        "dto.ClubChatCreatedDto": {
            "type": "object",
            "required": [
                "chat_id"
            ],
            "properties": {
                "chat_id": {
                    "type": "string"
                }
            }
        },
        "dto.ClubChatMessageDto": {
            "type": "object",
            "required": [
                "chat_id",
                "content",
                "created_at",
                "id",
                "sender_display_name",
                "sender_email",
                "sender_user_id"
            ],
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "sender_display_name": {
                    "type": "string"
                },
                "sender_email": {
                    "type": "string"
                },
                "sender_user_id": {
                    "type": "string"
                }
            }
        },
        "dto.ClubChatPostMessageDto": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "dto.ClubChatSummaryDto": {
            "type": "object",
            "required": [
                "chat_id",
                "club_id",
                "last_message",
                "last_message_at",
                "last_sender_user_id",
                "other_display_name",
                "other_email",
                "other_user_id"
            ],
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "club_id": {
                    "type": "string"
                },
                "last_message": {
                    "type": "string"
                },
                "last_message_at": {
                    "type": "string"
                },
                "last_sender_user_id": {
                    "type": "string"
                },
                "other_display_name": {
                    "type": "string"
                },
                "other_email": {
                    "type": "string"
                },
                "other_user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.ClubMessageCandidateDto": {
            "type": "object",
            "required": [
                "display_name",
                "email",
                "user_id"
            ],
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.ConcertDto": {
            "type": "object",
            "required": [
                "description",
                "dueDate",
                "hints",
                "id",
                "location",
                "noteInConcerts",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "hints": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "noteInConcerts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.NoteInConcertDto"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.ConcertPostDto": {
            "type": "object",
            "required": [
                "description",
                "dueDate",
                "hints",
                "location",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "hints": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.Event": {
            "type": "object",
            "required": [
                "status",
                "summary",
                "uid",
                "url"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "geoDateX": {
                    "type": "number"
                },
                "geoDateY": {
                    "type": "number"
                },
                "location": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                },
                "tzId": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dto.Folder": {
            "type": "object",
            "required": [
                "creationDate",
                "creator",
                "description",
                "id",
                "name",
                "type"
            ],
            "properties": {
                "creationDate": {
                    "type": "string"
                },
                "creator": {
                    "$ref": "#/definitions/dto.User"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "$ref": "#/definitions/dto.Folder"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "folder"
                    ]
                }
            }
        },
...
        "dto.Note": {
            "type": "object",
            "required": [
                "author",
                "creationDate",
                "creator",
                "description",
                "id",
                "name",
                "numberOfPages",
                "pdfAvailable",
                "type"
            ],
            "properties": {
                "author": {
                    "$ref": "#/definitions/dto.Author"
                },
                "creationDate": {
                    "type": "string"
                },
                "creator": {
                    "$ref": "#/definitions/dto.User"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "numberOfPages": {
                    "type": "integer"
                },
                "parent": {
                    "$ref": "#/definitions/dto.Folder"
                },
                "pdfAvailable": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "note"
                    ]
                }
            }
        },
...
        "dto.NoteInConcertDto": {
            "type": "object",
            "required": [
                "noteInConcert",
                "placeInConcert"
            ],
            "properties": {
                "noteInConcert": {
                    "$ref": "#/definitions/dto.Note"
                },
                "placeInConcert": {
                    "type": "integer"
                }
            }
        },
...
        "dto.User": {
            "type": "object",
            "required": [
                "email",
                "firstname",
                "lastname",
                "sideBarCollapsed",
                "telephoneNumber",
                "userId",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "profilePicUrl": {
                    "type": "string"
                },
                "sideBarCollapsed": {
                    "type": "boolean"
                },
                "telephoneNumber": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}